requirements.txt
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.20.0
sentencepiece
numpy
scipy
librosa
sounddevice
pyaudio
speechrecognition
gTTS
uvicorn
fastapi
pydantic
python-dotenv
huggingface_hub
requests
paho-mqtt
bitsandbytes
torchaudio